Initial Release
This is a working start to a GUI implementation of the children’s card
game War.  So far, the program includes a welcome screen with “play”
and “rules” buttons.  The “rules” button does not yet function.  The
”play” button leads to a new player screen which then leads to the main
game screen.  So far, the main screen only displays one face down card
button that still must be event handled.